Changed the abbreviated state name (NSW) to the full name (New South Wales) to be consistent with other holiday providers.
