+++ tests/Australia/NewSouthWales/EasterSundayTest.php
@@ -10,7 +10,7 @@
  * @author Sacha Telgenhof <me@sachatelgenhof.com>
  */
 
-namespace Yasumi\tests\Australia\NSW;
+namespace Yasumi\tests\Australia\NewSouthWales;
 
 use DateTime;
 use DateTimeZone;
@@ -20,17 +20,17 @@ use Yasumi\Holiday;
 use Yasumi\tests\YasumiTestCaseInterface;
 
 /**
- * Class for testing Bank Holiday in NSW (Australia)..
+ * Class for testing Easter Sunday in New South Wales (Australia)..
  */
-class BankHolidayTest extends NSWBaseTestCase implements YasumiTestCaseInterface
+class EasterSundayTest extends NewSouthWalesBaseTestCase implements YasumiTestCaseInterface
 {
     /**
      * The name of the holiday
      */
-    public const HOLIDAY = 'bankHoliday';
+    public const HOLIDAY = 'easter';
 
     /**
-     * Tests Bank Holiday
+     * Tests Easter Sunday
      *
      * @dataProvider HolidayDataProvider
      *
@@ -54,22 +54,20 @@ class BankHolidayTest extends NSWBaseTestCase implements YasumiTestCaseInterface
      * Returns a list of test dates
      *
      * @return array list of test dates for the holiday defined in this test
+     * @throws Exception
      */
     public function HolidayDataProvider(): array
     {
-        return [
-            [2010, '2010-08-02'],
-            [2011, '2011-08-01'],
-            [2012, '2012-08-06'],
-            [2013, '2013-08-05'],
-            [2014, '2014-08-04'],
-            [2015, '2015-08-03'],
-            [2016, '2016-08-01'],
-            [2017, '2017-08-07'],
-            [2018, '2018-08-06'],
-            [2019, '2019-08-05'],
-            [2020, '2020-08-03'],
-        ];
+        $data = [];
+
+        for ($y = 0; $y < 50; $y++) {
+            $year = $this->generateRandomYear();
+            $date = $this->calculateEaster($year, $this->timezone);
+
+            $data[] = [$year, $date->format('Y-m-d')];
+        }
+
+        return $data;
     }
 
     /**
@@ -81,8 +79,8 @@ class BankHolidayTest extends NSWBaseTestCase implements YasumiTestCaseInterface
         $this->assertTranslatedHolidayName(
             $this->region,
             self::HOLIDAY,
-            $this->generateRandomYear(1990),
-            [self::LOCALE => 'Bank Holiday']
+            $this->generateRandomYear(),
+            [self::LOCALE => 'Easter Sunday']
         );
     }
 
@@ -92,6 +90,6 @@ class BankHolidayTest extends NSWBaseTestCase implements YasumiTestCaseInterface
      */
     public function testHolidayType(): void
     {
-        $this->assertHolidayType($this->region, self::HOLIDAY, $this->generateRandomYear(1990), Holiday::TYPE_BANK);
+        $this->assertHolidayType($this->region, self::HOLIDAY, $this->generateRandomYear(), Holiday::TYPE_OFFICIAL);
     }
 }
